+++ src/pydiverse/pipedag/backend/table/sql/hooks.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# Copyright (c) QuantCo and pydiverse contributors 2025-2026
 # SPDX-License-Identifier: BSD-3-Clause
 
 import abc
@@ -1439,7 +1439,7 @@ class PolarsTableHook(DataframeSqlTableHook, TableHook[SQLTableStore]):
                 store.logger.exception(
                     f"Failed writing table using ADBC, falling back to sqlalchemy: {table.name}",
                 )
-                store.execute(DropTable(table, schema, if_exists=True, cascade=True))
+                store.execute(DropTable(table.name, schema, if_exists=True, cascade=True))
         df.write_database(
             f"{schema_name}.{table_name}",
             engine,