+++ docs/source/changelog.md
@@ -3,6 +3,7 @@
 ## 0.12.11 (2026-01-21)
 - Feat: simplify accessing tables in ParquetTableStore with polars
 - Fix: Fix mssqlkit bulk upload
+- Speed up pandas/polars transfer to/from duckdb in SqlTableStore
 
 ## 0.12.10 (2026-01-20)
 - Feat: Add ParquetTableStore.sync_metadata() function to sync local duckdb file with metadata store without running